requirements.txt
Django==1.10.5
django-filter==1.0.1
djangorestframework==3.5.4
djangorestframework-jwt==1.9.0
Markdown==2.6.8
mongoengine==0.11.0
PyJWT==1.4.2
pymongo==3.4.0
six==1.10.0
